Correctly cast `name_length`

+++ src/scanner.cc
@@ -71,7 +71,7 @@ struct Scanner {
         Tag &tag = tags[j];
         tag.type = static_cast<TagType>(buffer[i++]);
         if (tag.type == CUSTOM) {
-          uint16_t name_length = (uint16_t)buffer[i++];
+          uint16_t name_length = static_cast<uint8_t>(buffer[i++]);
           tag.custom_tag_name.assign(&buffer[i], &buffer[i + name_length]);
           i += name_length;
         }
